shared updated config
